uninstrusive multi-client changes
