+++ lib/client-multi.js
@@ -1,0 +1,34 @@
+/**
+ * This file is part of Shorty.
+ *
+ * Shorty is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; version 3 of the License.
+ *
+ * Shorty is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Shorty.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * @category   shorty
+ * @license    http://www.gnu.org/licenses/gpl-3.0.txt GPL
+ * @copyright  Copyright 2010 Evan Coury (http://www.Evan.pro/)
+ * @package    client
+ */
+
+var Connection  = require('./client').client,
+    util        = require('util');
+
+exports.multiClient = function(config, smppDefs) {
+  if (!config instanceof Array)
+    throw new Error('Multi-Client config is not an array');
+
+  for (var i = 0; i < config.length; i++) {
+    this.push(new Connection(config[i], smppDefs));
+  }
+
+};
+util.inherits(exports.multiClient, Array);
