+++ lib/shorty.js
@@ -20,6 +20,7 @@
  */
 
 var client = require('./client').client,
+    multiClient = require('./client-multi').multiClient,
     server = require('./server').server,
     fs     = require('fs'),
     smpp   = require('./smpp-definitions'),
@@ -51,6 +52,14 @@ exports.createClient = function(config) {
     }
 
     return new client(config.smpp, shorty.getSmppDefinitions());
+};
+
+exports.createMultiClient = function(config) {
+    if (typeof config === "string") {
+        config = exports.loadConfig(config);
+    }
+
+    return new multiClient(config.smpp, shorty.getSmppDefinitions());
 };
 
 exports.createServer = function(config) {
